refactor: Move Avatar components.
