+++ src/Components/AddressButton/AddressButton.tsx
@@ -1,7 +1,7 @@
 import { isDesktop } from 'react-device-detect';
 import React from 'react';
 import useENSAvatar from 'hooks/Guilds/ether-swr/ens/useENSAvatar';
-import Avatar from 'old-components/Guilds/Avatar';
+import Avatar from 'Components/Primitives/Avatar';
 import { shortenAddress } from 'utils';
 import { Badge } from 'old-components/Guilds/common/Badge';
 import { Loading } from 'Components/Primitives/Loading';
