+++ src/Components/Primitives/Avatar/index.tsx
@@ -1,0 +1,2 @@
+export * from "./Avatar";
+export { default } from "./Avatar";
